Adding my changes after my reset. Just some typos
