+++ resources/src/commands.rs
@@ -4,5 +4,6 @@
 // Read more here: https://doc.rust-lang.org/rust-by-example/mod.html
 
 pub mod buns;
+pub mod conversations_open;
 pub mod onboard_user;
 pub mod ping;
