+++ resources/src/slack.rs
@@ -1,4 +1,4 @@
-pub mod chat;
+pub mod add_reaction;
+pub mod chat_post_ephemeral_message;
+pub mod chat_post_message;
 pub mod client;
-pub mod conversations;
-pub mod reactions;
