+++ resources/src/slack/add_reaction.rs
@@ -1,14 +1,9 @@
-use slack_morphism::api::{SlackApiReactionsAddRequest, SlackApiReactionsAddResponse};
-use slack_morphism::hyper_tokio::SlackClientHyperConnector;
-use slack_morphism::{SlackApiToken, SlackClient};
+use slack_morphism::prelude::*;
 
 use crate::slack::client::build_token;
 
-/**
- * Adds the specified reaction to a message identified by its timestamp.
- * See https://api.slack.com/methods/reactions.add for request JSON examples.
- */
-pub async fn add(
+// This function will add the specified reaction to a message identified by its timestamp.
+pub async fn run(
     channel: &str,
     timestamp: &str,
     reaction: &str,
